Adapt kind.sh for multi-arch
Make it tolerate ktunnel pid gone

+++ kind.sh
@@ -4,20 +4,27 @@ cd $(dirname $0)
 
 : ${WORKSPACE_TMP:=/tmp}
 
+machine_name=$(uname -m)
+if [ "$machine_name" = "x86_64" ]
+then
+    arch=amd64
+else
+    arch=$machine_name
+fi
 export PATH=$WORKSPACE_TMP:$PATH
 if [ \! -x "$WORKSPACE_TMP/kind" ]
 then
-    curl -sLo "$WORKSPACE_TMP/kind" https://github.com/kubernetes-sigs/kind/releases/download/v0.23.0/kind-$(uname | tr '[:upper:]' '[:lower:]')-amd64
+    curl -sLo "$WORKSPACE_TMP/kind" https://github.com/kubernetes-sigs/kind/releases/download/v0.23.0/kind-$(uname | tr '[:upper:]' '[:lower:]')-${arch}
     chmod +x "$WORKSPACE_TMP/kind"
 fi
 if [ \! -x "$WORKSPACE_TMP/kubectl" ]
 then
-    curl -sLo "$WORKSPACE_TMP/kubectl" https://storage.googleapis.com/kubernetes-release/release/v1.30.1/bin/$(uname | tr '[:upper:]' '[:lower:]')/amd64/kubectl
+    curl -sLo "$WORKSPACE_TMP/kubectl" https://storage.googleapis.com/kubernetes-release/release/v1.30.1/bin/$(uname | tr '[:upper:]' '[:lower:]')/${arch}/kubectl
     chmod +x "$WORKSPACE_TMP/kubectl"
 fi
 if [ \! -x "$WORKSPACE_TMP/ktunnel" ]
 then
-    (cd "$WORKSPACE_TMP"; curl -sL https://github.com/omrikiei/ktunnel/releases/download/v1.6.1/ktunnel_1.6.1_Linux_x86_64.tar.gz | tar xvfz - ktunnel)
+    (cd "$WORKSPACE_TMP"; curl -sL https://github.com/omrikiei/ktunnel/releases/download/v1.6.1/ktunnel_1.6.1_$(uname -o)_${machine_name}.tar.gz | tar xvfz - ktunnel)
 fi
 
 export cluster=ci$RANDOM
@@ -25,7 +32,7 @@ export KUBECONFIG="$WORKSPACE_TMP/kubeconfig-$cluster"
 kind create cluster --name $cluster --wait 5m
 function cleanup() {
     set +e
-    if [ -v ktunnel_pid ]
+    if [ -v ktunnel_pid ] && ps -p $ktunnel_pid > /dev/null
     then
         kill $ktunnel_pid
     fi
